# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Ain Farba cultural heritage website
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ArticleCategory =
  (typeof ArticleCategory)[keyof typeof ArticleCategory];

export const ArticleCategory = {
  heritage: "heritage",
  history: "history",
  literature: "literature",
  thought: "thought",
  identity: "identity",
} as const;

export interface Article {
  id: number;
  title: string;
  summary: string;
  content: string;
  author: string;
  category: ArticleCategory;
  imageUrl?: string | null;
  publishedAt: string;
  createdAt: string;
}

export type CreateArticleBodyCategory =
  (typeof CreateArticleBodyCategory)[keyof typeof CreateArticleBodyCategory];

export const CreateArticleBodyCategory = {
  heritage: "heritage",
  history: "history",
  literature: "literature",
  thought: "thought",
  identity: "identity",
} as const;

export interface CreateArticleBody {
  title: string;
  summary: string;
  content: string;
  author: string;
  category: CreateArticleBodyCategory;
  imageUrl?: string | null;
  publishedAt: string;
}

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus];

export const EventStatus = {
  upcoming: "upcoming",
  past: "past",
} as const;

export interface Event {
  id: number;
  title: string;
  description: string;
  date: string;
  location: string;
  imageUrl?: string | null;
  status: EventStatus;
  createdAt: string;
}

export type CreateEventBodyStatus =
  (typeof CreateEventBodyStatus)[keyof typeof CreateEventBodyStatus];

export const CreateEventBodyStatus = {
  upcoming: "upcoming",
  past: "past",
} as const;

export interface CreateEventBody {
  title: string;
  description: string;
  date: string;
  location: string;
  imageUrl?: string | null;
  status: CreateEventBodyStatus;
}

export type NotableSpecialty =
  (typeof NotableSpecialty)[keyof typeof NotableSpecialty];

export const NotableSpecialty = {
  scholar: "scholar",
  jurist: "jurist",
  poet: "poet",
  writer: "writer",
  thinker: "thinker",
} as const;

export interface Notable {
  id: number;
  name: string;
  role: string;
  bio: string;
  specialty: NotableSpecialty;
  era?: string | null;
  imageUrl?: string | null;
  createdAt: string;
}

export type CreateNotableBodySpecialty =
  (typeof CreateNotableBodySpecialty)[keyof typeof CreateNotableBodySpecialty];

export const CreateNotableBodySpecialty = {
  scholar: "scholar",
  jurist: "jurist",
  poet: "poet",
  writer: "writer",
  thinker: "thinker",
} as const;

export interface CreateNotableBody {
  name: string;
  role: string;
  bio: string;
  specialty: CreateNotableBodySpecialty;
  era?: string | null;
  imageUrl?: string | null;
}

export type SummaryStatsArticlesByCategoryItem = {
  category: string;
  count: number;
};

export type SummaryStatsNotablesBySpecialtyItem = {
  specialty: string;
  count: number;
};

export interface SummaryStats {
  totalArticles: number;
  totalEvents: number;
  totalNotables: number;
  upcomingEvents: number;
  articlesByCategory: SummaryStatsArticlesByCategoryItem[];
  notablesBySpecialty: SummaryStatsNotablesBySpecialtyItem[];
}

export interface RecentActivity {
  recentArticles: Article[];
  recentEvents: Event[];
  recentNotables: Notable[];
}

export type ListArticlesParams = {
  category?: ListArticlesCategory;
  limit?: number;
  offset?: number;
};

export type ListArticlesCategory =
  (typeof ListArticlesCategory)[keyof typeof ListArticlesCategory];

export const ListArticlesCategory = {
  heritage: "heritage",
  history: "history",
  literature: "literature",
  thought: "thought",
  identity: "identity",
} as const;

export type ListEventsParams = {
  status?: ListEventsStatus;
  limit?: number;
  offset?: number;
};

export type ListEventsStatus =
  (typeof ListEventsStatus)[keyof typeof ListEventsStatus];

export const ListEventsStatus = {
  upcoming: "upcoming",
  past: "past",
} as const;

export type ListNotablesParams = {
  specialty?: ListNotablesSpecialty;
  limit?: number;
  offset?: number;
};

export type ListNotablesSpecialty =
  (typeof ListNotablesSpecialty)[keyof typeof ListNotablesSpecialty];

export const ListNotablesSpecialty = {
  scholar: "scholar",
  jurist: "jurist",
  poet: "poet",
  writer: "writer",
  thinker: "thinker",
} as const;